Subproject 'System Path Validators': Added inital implementation of 'spavalid.h' header with facilities for the user supplied system path validation.
Added initial implementation of validate_path()
* Implemented validation for empty path
* Implemented validation for existing path
* Implemented validation for directory path
* Implemented validation for empty directory

This commit includes the initial implementation of the 'validate_path()'
function. The following validations have been completed:
* Empty path: Check if the user-supplied path is empty.
* Existing path: Verify if the path provided by the user actually
exists.
* Driectory path: Determine if the path corresponds to a directory.
* Empty direcotory: Confirm if the direcotry is empty.

Pending tasks:
* Implement validation for file path and empty file.
* Complete implementation for the POSIX platform.

The upcoming tasks will focus on finaliying the validation logic for
file paths and empty files. Additionally, the code will be extended to
support POSIX platform. This commit serves as a checkpoint for the
completed validations and provides a foundation for further development.